backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Filipe",email = "fmartns66@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "requests (>=2.33.1,<3.0.0)",
    "wordfreq (>=3.1.1,<4.0.0)",
    "psycopg[binary] (>=3.2,<4)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
